refactor: move LanguageMenu component to context directory and update imports
